requirements.txt
# Minimal requirements for wave simulation
numpy>=1.24.0
matplotlib>=3.7.0

# Dataset generation requirements  
h5py>=3.8.0
tqdm>=4.65.0

# Deep learning requirements
torch>=2.0.0
torchvision>=0.15.0

# Experiment tracking and hyperparameter optimization
mlflow>=2.8.0
scikit-learn>=1.3.0
tensorboard>=2.15.0
wandb>=0.16.0

# Analysis and visualization requirements
scipy>=1.10.0
seaborn>=0.12.0
pandas>=2.0.0
torch-lucent>=0.1.8

# Model export and visualization
onnx>=1.18.0
torchviz>=0.0.2
graphviz>=0.20.0 